+++ docs-src/components/index.js
@@ -18,6 +18,7 @@ export { Tree } from './tree';
 
 export { Menu } from './menu';
 export { Icon } from './icon';
+export { Tag } from './tag';
 export { Utils } from './utils';
 export { Splitter } from './splitter';
 export { ColorPalette } from './color-palette';