+++ src/input/index.js
@@ -8,6 +8,7 @@ export * from './input-number';
 export * from './input-password';
 export * from './input-rating';
 export * from './input-search';
+export * from './input-tag';
 export * from './input-text';
 export * from './label';
 export * from './radio';