+++ src/utils.js
@@ -326,3 +326,13 @@ export function isInScrollable (node) {
 	}
 	return false;
 }
+
+
+export function isColorDark (hex) {
+	const color = (hex.charAt(0) === '#') ? hex.substring(1, 7) : hex;
+	const r = parseInt(color.substring(0, 2), 16); // red
+	const g = parseInt(color.substring(2, 4), 16); // green
+	const b = parseInt(color.substring(4, 6), 16); // blue
+	const brightness = ((r * 299) + (g * 587) + (b * 114)) / 1000;
+	return brightness < 140;
+}