+++ tests/helpers/Tag.svelte
@@ -1,0 +1,6 @@
+<Tag class="test-class" color="danger" icon="close" on:click>test text</Tag>
+
+<script>
+import { Tag } from '../../src/tag';
+
+</script>