+++ tests/input/InputDate.spec.js
@@ -2,8 +2,8 @@ import { render, fireEvent } from '@testing-library/svelte';
 import userEvent from '@testing-library/user-event';
 import jest from 'jest-mock';
 
-import { InputDate } from '../src/input/input-date';
-import { waitForTimeout } from './helpers/utils';
+import { InputDate } from '../../src/input/input-date';
+import { waitForTimeout } from '../helpers/utils';