+++ tests/input/InputMath.spec.js
@@ -1,8 +1,8 @@
 import { render } from '@testing-library/svelte';
 import userEvent from '@testing-library/user-event';
 import jest from 'jest-mock';
-import { InputMath } from '../src/input';
-import { waitForTimeout } from './helpers/utils';
+import { InputMath } from '../../src/input';
+import { waitForTimeout } from '../helpers/utils';
 
 
 const props = {