+++ tests/input/InputNumber.spec.js
@@ -1,7 +1,8 @@
 import { render } from '@testing-library/svelte';
 import jest from 'jest-mock';
-import { Textarea } from '../src/input';
-import { waitForTimeout } from './helpers/utils';
+import { InputNumber } from '../../src/input';
+import userEvent from '@testing-library/user-event';
+import { waitForTimeout } from '../helpers/utils';
 
 
 const props = {
@@ -16,8 +17,8 @@ const props = {
 };
 
 
-test('InputText', async () => {
-	const { container, component, getByTitle } = render(Textarea, props);
+test('InputNumber', async () => {
+	const { container, component, getByTitle } = render(InputNumber, props);
 	const mock = jest.fn();
 	component.$on('change', mock);
 
@@ -56,4 +57,42 @@ test('InputText', async () => {
 	expect(lbl).toBeInTheDocument();
 	expect(lbl).toHaveAttribute('for', props.id);
 	expect(lbl).toHaveTextContent(props.label);
+
+
+
+
+	// test just digits
+	await userEvent.clear(input);
+	await userEvent.type(input, '123456');
+	await userEvent.keyboard('[Tab]');
+	expect(input).toHaveValue('123456');
+
+	// test fractions
+	await userEvent.clear(input);
+	await userEvent.type(input, '12.3456');
+	await userEvent.keyboard('[Tab]');
+	expect(input).toHaveValue('12.3456');
+	expect(mock).toHaveBeenCalled();
+
+
+	// test incorrect input
+	await userEvent.clear(input);
+	await userEvent.type(input, '.0.0');
+	await userEvent.keyboard('[Tab]');
+	expect(input).toHaveValue('0');
+
+	await userEvent.clear(input);
+	await userEvent.type(input, 'abc');
+	await userEvent.keyboard('[Enter]');
+	expect(input).toHaveValue('');
+
+	await userEvent.clear(input);
+	await userEvent.type(input, '1.0000a');
+	await userEvent.keyboard('[Tab]');
+	expect(input).toHaveValue('1');
+
+	await userEvent.clear(input);
+	await userEvent.type(input, '1e');
+	await userEvent.keyboard('[Tab]');
+	expect(input).toHaveValue('1');
 });