+++ tests/input/InputPassword.spec.js
@@ -3,8 +3,8 @@ import userEvent from '@testing-library/user-event';
 import jest from 'jest-mock';
 import zxcvbn from 'zxcvbn';
 
-import { InputPassword } from '../src/input';
-import { waitForTimeout } from './helpers/utils';
+import { InputPassword } from '../../src/input';
+import { waitForTimeout } from '../helpers/utils';
 
 
 const props = {