+++ tests/input/Radio.spec.js
@@ -1,7 +1,7 @@
 import { render, fireEvent } from '@testing-library/svelte';
 import jest from 'jest-mock';
-import { Radio } from '../src/input';
-import { waitForTimeout } from './helpers/utils';
+import { Radio } from '../../src/input';
+import { waitForTimeout } from '../helpers/utils';
 
 
 test('Radio', async () => {