+++ tests/input/Toggle.spec.js
@@ -1,7 +1,7 @@
 import { render, fireEvent } from '@testing-library/svelte';
 import jest from 'jest-mock';
-import { Toggle } from '../src/input/toggle';
-import { waitForTimeout } from './helpers/utils';
+import { Toggle } from '../../src/input/toggle';
+import { waitForTimeout } from '../helpers/utils';
 
 
 test('Toggle', async () => {